+++ packages/react/astro.config.mjs
@@ -1,3 +1,0 @@
-import rootConfig from "../../astro.config.mjs"
-
-export default rootConfig