+++ packages/site/astro.config.mjs
@@ -1,0 +1,8 @@
+import rootConfig from "../../astro.config.mjs"
+
+export default {
+	...rootConfig,
+	buildOptions: {
+		site: "https://react.framework.dev",
+	},
+}