+++ packages/site/src/data/angular/libraries.ts
@@ -1,0 +1,5 @@
+import { Library } from "@framework/system/src/models/library"
+
+export const libraryTags = [] as const
+
+export const libraries: Library<typeof libraryTags[number]>[] = []